+++ monocle/main.py
@@ -212,6 +212,7 @@ def main():
                     updated_since=crawler_item['updated_since'],
                     loop_delay=tenant['crawler']['loop_delay'],
                     base_url=crawler_item['base_url'],
+                    insecure=crawler_item.get('insecure', False),
                     db=ELmonocleDB(
                         elastic_conn=args.elastic_conn,
                         index=tenant['index'],